started refactoring the architecture by moving stuff to free functions
