+++ azhal/src/vulkanHelper.h
@@ -1,12 +1,10 @@
 #pragma once
-
-#define  VK_LAYER_KHRONOS_VALIDATION_NAME "VK_LAYER_KHRONOS_validation"
+#include <vulkan/vulkan.hpp>
 
 namespace azhal
 {
-
 	template <typename T>
-	AZHAL_INLINE T CheckVkResultValue( const vk::ResultValue<T>& res_val, const AnsiChar* fail_msg )
+	AZHAL_INLINE T get_vk_result( const vk::ResultValue<T>& res_val, const AnsiChar* fail_msg = "" )
 	{
 		if( res_val.result != vk::Result::eSuccess )
 		{
@@ -18,7 +16,7 @@ namespace azhal
 	}
 
 	template <typename T>
-	AZHAL_INLINE T CheckVkResultValue( const vk::ResultValue<T>& res_val, const AnsiChar* fail_msg, std::initializer_list<vk::Result> successCodes )
+	AZHAL_INLINE T get_vk_result( const vk::ResultValue<T>& res_val, std::initializer_list<vk::Result> successCodes, const AnsiChar* fail_msg = "" )
 	{
 		if( std::find( successCodes.begin(), successCodes.end(), res_val.result ) != successCodes.end() )
 		{
@@ -28,5 +26,6 @@ namespace azhal
 
 		return res_val.value;
 	}
+
 
 }
